use new unified method to create annotations
