+++ src/main/java/de/marhali/easyi18n/editor/KeyAnnotator.java
@@ -1,11 +1,8 @@
 package de.marhali.easyi18n.editor;
 
 import com.intellij.lang.annotation.AnnotationHolder;
-import com.intellij.lang.annotation.Annotator;
 import com.intellij.lang.annotation.HighlightSeverity;
 import com.intellij.openapi.project.Project;
-import com.intellij.psi.PsiElement;
-import com.intellij.psi.PsiLiteralValue;
 
 import de.marhali.easyi18n.model.LocalizedNode;
 import de.marhali.easyi18n.service.DataStore;
@@ -14,28 +11,20 @@ import de.marhali.easyi18n.service.SettingsService;
 import org.jetbrains.annotations.NotNull;
 
 /**
- * Translation key annotator.
+ *
  * @author marhali
  */
-public class I18nKeyAnnotator implements Annotator {
+public class KeyAnnotator {
 
-    @Override
-    public void annotate(@NotNull PsiElement element, @NotNull AnnotationHolder holder) {
-        if(!(element instanceof PsiLiteralValue)) {
-            return;
-        }
-
-        PsiLiteralValue literalValue = (PsiLiteralValue) element;
-        String value = literalValue.getValue() instanceof String ? (String) literalValue.getValue() : null;
-
-        if(value == null) {
-            return;
-        }
-
-        Project project = element.getProject();
+    /**
+     * Adds annotations for i18n keys with content preview for preferred locale.
+     * @param key I18n key extracted by psi element
+     * @param project Project instance
+     * @param holder Annotation holder
+     */
+    protected void annotate(@NotNull String key, @NotNull Project project, @NotNull AnnotationHolder holder) {
         String previewLocale = SettingsService.getInstance(project).getState().getPreviewLocale();
-
-        LocalizedNode node = DataStore.getInstance(project).getTranslations().getNode(value);
+        LocalizedNode node = DataStore.getInstance(project).getTranslations().getNode(key);
 
         if(node == null) { // Unknown translation. Just ignore it
             return;
