+++ src/main/java/de/marhali/easyi18n/editor/generic/GenericKeyAnnotator.java
@@ -1,26 +1,28 @@
-package de.marhali.easyi18n.editor.kotlin;
+package de.marhali.easyi18n.editor.generic;
 
 import com.intellij.lang.annotation.AnnotationHolder;
 import com.intellij.lang.annotation.Annotator;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiLiteralValue;
 
 import de.marhali.easyi18n.editor.KeyAnnotator;
+
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.kotlin.psi.KtLiteralStringTemplateEntry;
 
 /**
- * Kotlin specific translation key annotator
+ * Translation key annotator.
  * @author marhali
  */
-public class KotlinKeyAnnotator extends KeyAnnotator implements Annotator {
+public class GenericKeyAnnotator extends KeyAnnotator implements Annotator {
 
     @Override
     public void annotate(@NotNull PsiElement element, @NotNull AnnotationHolder holder) {
-        if(!(element instanceof KtLiteralStringTemplateEntry)) {
+        if(!(element instanceof PsiLiteralValue)) {
             return;
         }
 
-        String value = element.getText();
+        PsiLiteralValue literalValue = (PsiLiteralValue) element;
+        String value = literalValue.getValue() instanceof String ? (String) literalValue.getValue() : null;
 
         if(value == null) {
             return;
