+++ src/main/resources/META-INF/plugin.xml
@@ -16,6 +16,6 @@
         <completion.contributor language="any"
                                 implementationClass="de.marhali.easyi18n.editor.I18nCompletionContributor" />
 
-        <annotator language="" implementationClass="de.marhali.easyi18n.editor.I18nKeyAnnotator" />
+        <annotator language="" implementationClass="de.marhali.easyi18n.editor.generic.GenericKeyAnnotator" />
     </extensions>
 </idea-plugin>
